% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edec_stage_1.R
\name{estimate_meth_qp}
\alias{estimate_meth_qp}
\title{Estimate cell type specific methylation}
\usage{
estimate_meth_qp(meth_bulk_samples, cell_type_props)
}
\arguments{
\item{meth_bulk_samples}{Matrix of methylation profiles of bulk complex
tissue samples. Columns correspond to different samples and rows correspond
to different loci/probes.}

\item{cell_type_props}{Matrix of proportions of constituent cell types.
Columns correspond to different cell types and rows correspond to different
bulk tissue samples.}
}
\value{
Matrix with estimated average methylation profiles of constituent
  cell types.
}
\description{
Given methylation profiles of complex tissue samples and propotions of
contituent cell types in each sample, \code{estimate_meth_qp} will estimate
average methylation profiles of constituent cell types across the set of
input samples.
}
\details{
EDec assumes that the methylation profiles of complex tissue samples
correspond to the linear combination of cell type proportions and methylation
profiles of each cell type. Given the methylation profiles of a set of
complex tissue samples and the proportions of constituent cell types in each
sample, this function estimates average methylation profiles of constituent
cell types by solving constrained least squares problems through quadratic
programming. The constraint is that the methylation profiles of constituent
cell types are numbers in the [0,1] interval.
}

