% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDec.R
\name{estimate_props_qp}
\alias{estimate_props_qp}
\title{Estimate cell type proportions}
\usage{
estimate_props_qp(meth_bulk_samples, cell_type_specific_meth)
}
\arguments{
\item{meth_bulk_samples}{Matrix of methylation profiles of bulk complex
tissue samples. Columns correspond to different samples and rows correspond
to different loci/probes.}

\item{cell_type_specific_meth}{Matrix of methylation profiles of constituent
cell types. Columns correspond to different cell types and rows correspond
to different loci/probes.}
}
\description{
This function will estimate the proportions of constituent cell types in
each input sample, given methylation profiles of complex tissue samples and
methylation profiles of constituent cell types.
}
\details{
EDec assumes that the methylation profiles of complex tissue samples
correspond to the linear combination of cell type proportions and
methylation profiles of each cell type. Given the methylation profiles of a
set of complex tissue samples and the methylation profiles of constituent
cell types this function estimates cell type proportions in each sample by
solving constrained least squares problems through quadratic programming. The
constraints are that the proportions of constituent cell types are numbers in
the [0,1] interval and that the proportions of all cell types in each sample
sum up to one.
}

