% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edec_aux_functions.R
\name{find_best_perm}
\alias{find_best_perm}
\title{Find best permutation of correlation matrix}
\usage{
find_best_perm(cor_matrix)
}
\arguments{
\item{cor_matrix}{A matrix of pairwise correlations between a set of numeric
vectors}
}
\value{
A list containing the following components:

\describe{
 \item{\code{pmatrix}}{The permuted correlation matrix.}
 \item{\code{pvec}}{The permutation vector, such that \code{pmatrix =
  cor_matrix[pvec, ]}.}
}
}
\description{
This function finds the permutation of a correlation matrix that is most
similar to the identity matrix.
}

