% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edec_stage_0.R
\name{perform_t_tests_all_classes_each_pair}
\alias{perform_t_tests_all_classes_each_pair}
\title{Comparison of each pair of classes}
\usage{
perform_t_tests_all_classes_each_pair(data_matrix, class_vector)
}
\arguments{
\item{data_matrix}{Matrix containing the data for all samples. Columns
correspond to different samples.}

\item{class_vector}{A vector of numbers or strings representing the classes
associated with each column in \code{data_matrix}.}
}
\value{
A list containing:

\describe{
  \item{\code{p.values}}{A matrix containing the p-values
    for each comparison. Each column corresponds to the comparison between a
    pair of classes.
  }
  \item{\code{diff.means}}{A matrix containing the
    difference between group means for each comparison. Each column
    corresponds to the comparison between a pair of classes.
  }
}
}
\description{
For each row of a matrix, this function performs T-tests comparing the values
of that row for samples in one class against the values in that row samples
in another class. Such comparisons are done for each pair of reference
classes.
}

